jina_2/audio-search_V1/requirements.txt
git+https://github.com/jina-ai/jina
resampy==0.2.2
tensorflow_cpu==2.2.0
tf_slim==1.1.0
six==1.15.0
soundfile==0.10.3.post1
librosa==0.8.0
click==7.1.2
pandas==1.2.2
